evdev: correct/revert change done in "021f312 - Correct X, Y axis not inverting when EVDEV_XXX_MIN > EVDEV_XXX_MAX (#87)"
The aim of the change was to activate EVDEV_CALIBRATE when EVDEV_SCALE was also in use. However, this ruined EVDEV_SCALE when EVDEV_CALIBRATE was NOT used.

EVDEV_SCALE is a simplified version of EVDEV_CALIBRATE, so these are exclusive. Both at the same time does not make any sense. So enabling EVDEV_SCALE always did override EVDEV_CALIBRATE.

This commit restores the code as it was before 021f312, but using the #elif instead of nested #if makes it more clear that it's either EVDEV_SCALE **or** EVDEV_CALIBRATE **or** no transformation at all.